change int type for float type

+++ app/models/account.py
@@ -4,7 +4,7 @@ from app.models.statement import StatementModelList
 
 class AccountModel:
     def __init__(self):
-        self._balance = 0
+        self._balance = 0.0
         self.agency = AgencyModel()
         self._statements = StatementModelList()
     
@@ -12,16 +12,16 @@ class AccountModel:
     def balance(self):
         return self._balance
     
-    def deposit(self, value: int, description: str = None):
+    def deposit(self, value: float, description: str = None):
         self._balance += value
         self._statements.add(balance=self.balance, description=description, deposit=value)
     
-    def withdraw(self, value: int, description: str = None):
+    def withdraw(self, value: float, description: str = None):
         self._balance -= value
         self._statements.add(balance=self.balance, description=description, withdraw=value)
         return value
 
-    def transfer(self, value: int, account: 'AccountModel', description: str = None):
+    def transfer(self, value: float, account: 'AccountModel', description: str = None):
         account.deposit(value, description)
         self.withdraw(value, description)
